phase2: merge files
