--- src/permlink.rs
@@ -1,7 +1,0 @@
-use compile;
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct Result {
-  parameter: compile::Parameter,
-  result: compile::Result,
-}
